feat(web): show folder colors on the manage documents page
